Code/Interface graphique/requirements.txt
ingescape>=4
PyQt5>=5.15.0
pywebview>=4.0
